requirements.txt
Django==1.7.4
Pillow==2.7.0
argparse==1.2.1
distribute==0.6.24
django-extensions==1.5.0
django-froala-editor==1.2.5
django-pure-pagination==0.2.1
eveapi==1.2.6
psycopg2==2.5.4
python-memcached==1.53
six==1.9.0
wsgiref==0.1.2
